#136 공통 - 예외 처리
