+++ src/main/java/com/gabia/gyebalja/common/exception/NotExistDataException.java
@@ -1,0 +1,11 @@
+package com.gabia.gyebalja.common.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.FORBIDDEN)
+public class NotExistDataException extends RuntimeException {
+    public NotExistDataException(String msg) {
+        super(msg);
+    }
+}
